Manually provide a Content-Type header for document requests
WP Document Revisions attempts to set a Content-Type header, but
WordPress sends the full set of headers to the browser before the
request is modified.

We hook in early here to provide Content-Type based on what we
know of the document from the save process. In the future it would
be nice to provide other header information as well.