Revert use of wsu-search as an ES index, use wsu-web
In d14bf25f98d12571, I filtered the search index to use a new
wsu-search instead of the default wsu-web, but failed to write
any kind of commit message explaining the decision. Here we are
months later and I'm trying to sort out *why*.

It appears that I did this so that the schema could be changed to
be more filterable. Unfortunately, I picked a crappy name for the
schema as this is all very WordPress specific and won't work for
WSU search as a whole.

Let's revert to using wsu-web for now. We can redefine the schema
attached to that and reindex sites using WP-CLI pretty easily.

This will free us up to use the wsu-search schema for an actual
WSU search interface.